requirements.txt
argparse
dash==2.10.0
matplotlib==3.7.1
numpy==1.24.3
openai==0.28
pandas==1.5.3
plotly==5.15.0
scikit-learn==1.2.2
torch==2.0.1
transformers==4.31.0
python-dotenv

